fix: Update IPv6 counter field names in wd_util.c
Update the IPv6 counter field names from 'incoming' and 'outgoing' to
'incoming_bytes' and 'outgoing_bytes' in wd_util.c to match the actual
structure definition in the traffic_stats structure.

This fixes compile errors in the wd_util.c file where it was trying
to access non-existent field names in the counters6 structure.

Signed-off-by: Dengfeng Liu <liudf0716@gmail.com>